WEB-2039 extract some native sheet parts
