+++ src/sheet-native.tsx
@@ -1,0 +1,186 @@
+import type {
+    NativeSheetImplementation,
+    SheetPropsByType,
+    SheetResultByType,
+    SheetType,
+    SheetTypeWithProps,
+    SheetTypeWithPropsUnion,
+} from './sheet-types';
+
+let nativeSheetImplementation: NativeSheetImplementation | null = null;
+
+export const setNativeSheetImplementation = (implementation: NativeSheetImplementation): void => {
+    nativeSheetImplementation = implementation;
+};
+
+const normalizeDescriptionForNative = (description?: string | Array<string>): string | undefined => {
+    if (Array.isArray(description)) {
+        if (description.length) {
+            return description.join('\n\n');
+        } else {
+            return undefined;
+        }
+    }
+    return description;
+};
+
+const showRadioListNativeSheet = ({
+    title,
+    subtitle,
+    description,
+    selectedId,
+    items,
+}: SheetPropsByType['RADIO_LIST']): Promise<SheetResultByType['RADIO_LIST']> => {
+    return (nativeSheetImplementation as NativeSheetImplementation)({
+        title,
+        subtitle,
+        // TODO: add multiline support to native sheet
+        description: normalizeDescriptionForNative(description),
+        content: [
+            {
+                type: 'LIST',
+                id: 'list-0',
+                listType: 'SINGLE_SELECTION',
+                autoSubmit: true,
+                selectedIds: typeof selectedId === 'string' ? [selectedId] : [],
+                items,
+            },
+        ],
+    }).then(({action, result}) => {
+        if (action === 'SUBMIT') {
+            return {
+                action,
+                selectedId: result[0].selectedIds[0],
+            };
+        } else {
+            return {
+                action,
+                selectedId: null,
+            };
+        }
+    });
+};
+
+const showActionsListNativeSheet = ({
+    title,
+    subtitle,
+    description,
+    items,
+}: SheetPropsByType['ACTIONS_LIST']): Promise<SheetResultByType['ACTIONS_LIST']> => {
+    return (nativeSheetImplementation as NativeSheetImplementation)({
+        title,
+        subtitle,
+        // TODO: add multiline support to native sheet
+        description: normalizeDescriptionForNative(description),
+        content: [
+            {
+                type: 'LIST',
+                id: 'list-0',
+                listType: 'ACTIONS',
+                autoSubmit: true,
+                selectedIds: [],
+                items,
+            },
+        ],
+    }).then(({action, result}) => {
+        if (action === 'SUBMIT') {
+            return {
+                action,
+                selectedId: result[0].selectedIds[0],
+            };
+        } else {
+            return {
+                action,
+                selectedId: null,
+            };
+        }
+    });
+};
+
+const showInfoNativeSheet = async ({
+    title,
+    subtitle,
+    description,
+    items,
+}: SheetPropsByType['INFO']): Promise<SheetResultByType['INFO']> => {
+    // nothing to return, this is an informative sheet
+    await (nativeSheetImplementation as NativeSheetImplementation)({
+        title,
+        subtitle,
+        // TODO: add multiline support to native sheet
+        description: normalizeDescriptionForNative(description),
+        content: [
+            {
+                type: 'LIST',
+                id: 'list-0',
+                listType: 'INFORMATIVE',
+                autoSubmit: false,
+                selectedIds: [],
+                items,
+            },
+        ],
+    });
+};
+
+const showActionsNativeSheet = ({
+    title,
+    subtitle,
+    description,
+    button,
+    secondaryButton,
+    link,
+}: SheetPropsByType['ACTIONS']): Promise<SheetResultByType['ACTIONS']> => {
+    return (nativeSheetImplementation as NativeSheetImplementation)({
+        title,
+        subtitle,
+        // TODO: add multiline support to native sheet
+        description: normalizeDescriptionForNative(description),
+        content: [
+            {
+                type: 'BOTTOM_ACTIONS',
+                id: 'bottom-actions-0',
+                button,
+                secondaryButton,
+                link,
+            },
+        ],
+    }).then(({action, result}) => {
+        if (action === 'SUBMIT') {
+            const bottomActionsResult = result.find(({id}) => id === 'bottom-actions-0');
+            const pressedAction = bottomActionsResult?.selectedIds[0];
+            if (pressedAction === 'PRIMARY' || pressedAction === 'SECONDARY' || pressedAction === 'LINK') {
+                return {
+                    action: pressedAction,
+                };
+            }
+        }
+        return {
+            action: 'DISMISS',
+        };
+    });
+};
+
+export const showNativeSheet = <T extends SheetType>(
+    sheetProps: SheetTypeWithProps<T>
+): Promise<SheetResultByType[T]> => {
+    let nativeResponse: Promise<SheetResultByType[T]>;
+    const {type, props} = sheetProps as SheetTypeWithPropsUnion;
+    switch (type) {
+        case 'INFO':
+            nativeResponse = showInfoNativeSheet(props) as Promise<SheetResultByType[T]>;
+            break;
+        case 'ACTIONS_LIST':
+            nativeResponse = showActionsListNativeSheet(props) as Promise<SheetResultByType[T]>;
+            break;
+        case 'RADIO_LIST':
+            nativeResponse = showRadioListNativeSheet(props) as Promise<SheetResultByType[T]>;
+            break;
+        case 'ACTIONS':
+            nativeResponse = showActionsNativeSheet(props) as Promise<SheetResultByType[T]>;
+            break;
+        default:
+            const unknownType: never = type;
+            throw new Error(`Unknown sheet type: ${unknownType}`);
+    }
+    return nativeResponse;
+};
